backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
scikit-learn==1.3.2
pandas==2.1.4
faker==20.1.0
langgraph==0.0.26
langchain==0.1.11
langchain-google-genai
python-dotenv
pycryptodome==3.19.0
liboqs-python==0.8.0
websockets==12.0
gunicorn==21.2.0
